+++ pokegym/data.py
@@ -501,63 +501,183 @@ items_dict = {
 }
 
 poke_dict = {
-   3: {'hex': '3', 'decimal': '3', 'name': 'Nidoran♂'},
-   4: {'hex': '4', 'decimal': '4', 'name': 'Clefairy'},
-   5: {'hex': '5', 'decimal': '5', 'name': 'Spearow'},
-   9: {'hex': '9', 'decimal': '9', 'name': 'Ivysaur'},
-   15: {'hex': 'F', 'decimal': '15', 'name': 'Nidoran♀'},
-   16: {'hex': '10', 'decimal': '16', 'name': 'Nidoqueen'},
-   28: {'hex': '1C', 'decimal': '28', 'name': 'Blastoise'},
-   35: {'hex': '23', 'decimal': '35', 'name': 'Fearow'},
-   36: {'hex': '24', 'decimal': '36', 'name': 'Pidgey'},
-   38: {'hex': '26', 'decimal': '38', 'name': 'Kadabra'},
-   39: {'hex': '27', 'decimal': '39', 'name': 'Graveler'},
-   46: {'hex': '2E', 'decimal': '46', 'name': 'Parasect', 'type': 'Bug'},
-   48: {'hex': '30', 'decimal': '48', 'name': 'Drowzee'},
-   49: {'hex': '31', 'decimal': '49', 'name': 'Golem'},
-   57: {'hex': '39', 'decimal': '57', 'name': 'Mankey'},
-   59: {'hex': '3B', 'decimal': '59', 'name': 'Diglett'},
-   70: {'hex': '46', 'decimal': '70', 'name': 'Doduo'},
-   84: {'hex': '54', 'decimal': '84', 'name': 'Pikachu', 'type': 'Electric'},
-   85: {'hex': '55', 'decimal': '85', 'name': 'Raichu', 'type': 'Electric'},
-   100: {'hex': '64', 'decimal': '100', 'name': 'Jigglypuff'},
-   101: {'hex': '65', 'decimal': '101', 'name': 'Wigglytuff'},
-   107: {'hex': '6B', 'decimal': '107', 'name': 'Zubat'},
-   108: {'hex': '6C', 'decimal': '108', 'name': 'Ekans'},
-   109: {'hex': '6D', 'decimal': '109', 'name': 'Paras', 'type': 'Bug'},
-   112: {'hex': '70', 'decimal': '112', 'name': 'Weedle', 'type': 'Bug'},
-   113: {'hex': '71', 'decimal': '113', 'name': 'Kakuna', 'type': 'Bug'},
-   114: {'hex': '72', 'decimal': '114', 'name': 'Beedrill', 'type': 'Bug'},
-   116: {'hex': '74', 'decimal': '116', 'name': 'Dodrio'},
-   117: {'hex': '75', 'decimal': '117', 'name': 'Primeape'},
-   118: {'hex': '76', 'decimal': '118', 'name': 'Dugtrio'},
-   123: {'hex': '7B', 'decimal': '123', 'name': 'Caterpie', 'type': 'Bug'},
-   124: {'hex': '7C', 'decimal': '124', 'name': 'Metapod', 'type': 'Bug'},
-   125: {'hex': '7D', 'decimal': '125', 'name': 'Butterfree', 'type': 'Bug'},
-   129: {'hex': '81', 'decimal': '129', 'name': 'Hypno'},
-   130: {'hex': '82', 'decimal': '130', 'name': 'Golbat'},
-   133: {'hex': '85', 'decimal': '133', 'name': 'Magikarp'},
-   142: {'hex': '8E', 'decimal': '142', 'name': 'Clefable'},
-   148: {'hex': '94', 'decimal': '148', 'name': 'Abra'},
-   149: {'hex': '95', 'decimal': '149', 'name': 'Alakazam'},
-   150: {'hex': '96', 'decimal': '150', 'name': 'Pidgeotto'},
-   151: {'hex': '97', 'decimal': '151', 'name': 'Pidgeot'},
-   153: {'hex': '99', 'decimal': '153', 'name': 'Bulbasaur'},
-   154: {'hex': '9A', 'decimal': '154', 'name': 'Venusaur'},
-   165: {'hex': 'A5', 'decimal': '165', 'name': 'Rattata'},
-   166: {'hex': 'A6', 'decimal': '166', 'name': 'Raticate'},
-   167: {'hex': 'A7', 'decimal': '167', 'name': 'Nidorino'},
-   168: {'hex': 'A8', 'decimal': '168', 'name': 'Nidorina'},
-   169: {'hex': 'A9', 'decimal': '169', 'name': 'Geodude'},
-   176: {'hex': 'B0', 'decimal': '176', 'name': 'Charmander'},
-   177: {'hex': 'B1', 'decimal': '177', 'name': 'Squirtle'},
-   178: {'hex': 'B2', 'decimal': '178', 'name': 'Charmeleon'},
-   179: {'hex': 'B3', 'decimal': '179', 'name': 'Wartortle'},
-   180: {'hex': 'B4', 'decimal': '180', 'name': 'Charizard'},
-   185: {'hex': 'B9', 'decimal': '185', 'name': 'Oddish'},
-   186: {'hex': 'BA', 'decimal': '186', 'name': 'Gloom'},
-   187: {'hex': 'BB', 'decimal': '187', 'name': 'Vileplume'}
-}
+    1: {'hex': '1', 'decimal': '1', 'name': 'Rhydon'},
+    2: {'hex': '2', 'decimal': '2', 'name': 'Kangaskhan'},
+    3: {'hex': '3', 'decimal': '3', 'name': 'Nidoran♂'},
+    4: {'hex': '4', 'decimal': '4', 'name': 'Clefairy'},
+    5: {'hex': '5', 'decimal': '5', 'name': 'Spearow'},
+    6: {'hex': '6', 'decimal': '6', 'name': 'Voltorb', 'type': 'Electric'},
+    7: {'hex': '7', 'decimal': '7', 'name': 'Nidoking'},
+    8: {'hex': '8', 'decimal': '8', 'name': 'Slowbro'},
+    9: {'hex': '9', 'decimal': '9', 'name': 'Ivysaur'},
+    10: {'hex': 'A', 'decimal': '10', 'name': 'Exeggutor'},
+    11: {'hex': 'B', 'decimal': '11', 'name': 'Lickitung'},
+    12: {'hex': 'C', 'decimal': '12', 'name': 'Exeggcute'},
+    13: {'hex': 'D', 'decimal': '13', 'name': 'Grimer'},
+    14: {'hex': 'E', 'decimal': '14', 'name': 'Gengar', 'type': 'Ghost'},
+    15: {'hex': 'F', 'decimal': '15', 'name': 'Nidoran♀'},
+    16: {'hex': '10', 'decimal': '16', 'name': 'Nidoqueen'},
+    17: {'hex': '11', 'decimal': '17', 'name': 'Cubone'},
+    18: {'hex': '12', 'decimal': '18', 'name': 'Rhyhorn'},
+    19: {'hex': '13', 'decimal': '19', 'name': 'Lapras', 'type': 'Ice'},
+    20: {'hex': '14', 'decimal': '20', 'name': 'Arcanine'},
+    21: {'hex': '15', 'decimal': '21', 'name': 'Mew'},
+    22: {'hex': '16', 'decimal': '22', 'name': 'Gyarados'},
+    23: {'hex': '17', 'decimal': '23', 'name': 'Shellder'},
+    24: {'hex': '18', 'decimal': '24', 'name': 'Tentacool'},
+    25: {'hex': '19', 'decimal': '25', 'name': 'Gastly', 'type': 'Ghost'},
+    26: {'hex': '1A', 'decimal': '26', 'name': 'Scyther', 'type': 'Bug'},
+    27: {'hex': '1B', 'decimal': '27', 'name': 'Staryu'},
+    28: {'hex': '1C', 'decimal': '28', 'name': 'Blastoise'},
+    29: {'hex': '1D', 'decimal': '29', 'name': 'Pinsir', 'type': 'Bug'},
+    30: {'hex': '1E', 'decimal': '30', 'name': 'Tangela'},
+    31: {'hex': '1F', 'decimal': '31', 'name': 'MissingNo. (Scizor)'},
+    32: {'hex': '20', 'decimal': '32', 'name': 'MissingNo. (Shuckle)'},
+    33: {'hex': '21', 'decimal': '33', 'name': 'Growlithe'},
+    34: {'hex': '22', 'decimal': '34', 'name': 'Onix'},
+    35: {'hex': '23', 'decimal': '35', 'name': 'Fearow'},
+    36: {'hex': '24', 'decimal': '36', 'name': 'Pidgey'},
+    37: {'hex': '25', 'decimal': '37', 'name': 'Slowpoke'},
+    38: {'hex': '26', 'decimal': '38', 'name': 'Kadabra'},
+    39: {'hex': '27', 'decimal': '39', 'name': 'Graveler'},
+    40: {'hex': '28', 'decimal': '40', 'name': 'Chansey'},
+    41: {'hex': '29', 'decimal': '41', 'name': 'Machoke'},
+    42: {'hex': '2A', 'decimal': '42', 'name': 'Mr. Mime'},
+    43: {'hex': '2B', 'decimal': '43', 'name': 'Hitmonlee'},
+    44: {'hex': '2C', 'decimal': '44', 'name': 'Hitmonchan'},
+    45: {'hex': '2D', 'decimal': '45', 'name': 'Arbok'},
+    46: {'hex': '2E', 'decimal': '46', 'name': 'Parasect', 'type': 'Bug'},
+    47: {'hex': '2F', 'decimal': '47', 'name': 'Psyduck'},
+    48: {'hex': '30', 'decimal': '48', 'name': 'Drowzee'},
+    49: {'hex': '31', 'decimal': '49', 'name': 'Golem'},
+    50: {'hex': '32', 'decimal': '50', 'name': 'MissingNo. (Heracross)'},
+    51: {'hex': '33', 'decimal': '51', 'name': 'Magmar'},
+    52: {'hex': '34', 'decimal': '52', 'name': 'MissingNo. (Ho-Oh)'},
+    53: {'hex': '35', 'decimal': '53', 'name': 'Electabuzz', 'type': 'Electric'},
+    54: {'hex': '36', 'decimal': '54', 'name': 'Magneton', 'type': 'Electric'},
+    55: {'hex': '37', 'decimal': '55', 'name': 'Koffing'},
+    56: {'hex': '38', 'decimal': '56', 'name': 'MissingNo. (Sneasel)'},
+    57: {'hex': '39', 'decimal': '57', 'name': 'Mankey'},
+    58: {'hex': '3A', 'decimal': '58', 'name': 'Seel'},
+    59: {'hex': '3B', 'decimal': '59', 'name': 'Diglett'},
+    60: {'hex': '3C', 'decimal': '60', 'name': 'Tauros'},
+    61: {'hex': '3D', 'decimal': '61', 'name': 'MissingNo. (Teddiursa)'},
+    62: {'hex': '3E', 'decimal': '62', 'name': 'MissingNo. (Ursaring)'},
+    63: {'hex': '3F', 'decimal': '63', 'name': 'MissingNo. (Slugma)'},
+    64: {'hex': '40', 'decimal': '64', 'name': "Farfetch'd"},
+    65: {'hex': '41', 'decimal': '65', 'name': 'Venonat', 'type': 'Bug'},
+    66: {'hex': '42', 'decimal': '66', 'name': 'Dragonite', 'type': 'Dragon'},
+    67: {'hex': '43', 'decimal': '67', 'name': 'MissingNo. (Magcargo)'},
+    68: {'hex': '44', 'decimal': '68', 'name': 'MissingNo. (Swinub)'},
+    69: {'hex': '45', 'decimal': '69', 'name': 'MissingNo. (Piloswine)'},
+    70: {'hex': '46', 'decimal': '70', 'name': 'Doduo'},
+    71: {'hex': '47', 'decimal': '71', 'name': 'Poliwag'},
+    72: {'hex': '48', 'decimal': '72', 'name': 'Jynx', 'type': 'Ice'},
+    73: {'hex': '49', 'decimal': '73', 'name': 'Moltres'},
+    74: {'hex': '4A', 'decimal': '74', 'name': 'Articuno', 'type': 'Ice'},
+    75: {'hex': '4B', 'decimal': '75', 'name': 'Zapdos', 'type': 'Electric'},
+    76: {'hex': '4C', 'decimal': '76', 'name': 'Ditto'},
+    77: {'hex': '4D', 'decimal': '77', 'name': 'Meowth'},
+    78: {'hex': '4E', 'decimal': '78', 'name': 'Krabby'},
+    79: {'hex': '4F', 'decimal': '79', 'name': 'MissingNo. (Corsola)'},
+    80: {'hex': '50', 'decimal': '80', 'name': 'MissingNo. (Remoraid)'},
+    81: {'hex': '51', 'decimal': '81', 'name': 'MissingNo. (Octillery)'},
+    82: {'hex': '52', 'decimal': '82', 'name': 'Vulpix'},
+    83: {'hex': '53', 'decimal': '83', 'name': 'Ninetales'},
+    84: {'hex': '54', 'decimal': '84', 'name': 'Pikachu', 'type': 'Electric'},
+    85: {'hex': '55', 'decimal': '85', 'name': 'Raichu', 'type': 'Electric'},
+    86: {'hex': '56', 'decimal': '86', 'name': 'MissingNo. (Deli)'},
+    87: {'hex': '57', 'decimal': '87', 'name': 'MissingNo. (Mantine)'},
+    88: {'hex': '58', 'decimal': '88', 'name': 'Dratini', 'type': 'Dragon'},
+    89: {'hex': '59', 'decimal': '89', 'name': 'Dragonair', 'type': 'Dragon'},
+    90: {'hex': '5A', 'decimal': '90', 'name': 'Kabuto'},
+    91: {'hex': '5B', 'decimal': '91', 'name': 'Kabutops'},
+    92: {'hex': '5C', 'decimal': '92', 'name': 'Horsea'},
+    93: {'hex': '5D', 'decimal': '93', 'name': 'Seadra'},
+    94: {'hex': '5E', 'decimal': '94', 'name': 'MissingNo. (Skarmory)'},
+    95: {'hex': '5F', 'decimal': '95', 'name': 'MissingNo. (Houndour)'},
+    96: {'hex': '60', 'decimal': '96', 'name': 'Sandshrew'},
+    97: {'hex': '61', 'decimal': '97', 'name': 'Sandslash'},
+    98: {'hex': '62', 'decimal': '98', 'name': 'Omanyte'},
+    99: {'hex': '63', 'decimal': '99', 'name': 'Omastar'},
+    100: {'hex': '64', 'decimal': '100', 'name': 'Jigglypuff'},
+    101: {'hex': '65', 'decimal': '101', 'name': 'Wigglytuff'},
+    102: {'hex': '66', 'decimal': '102', 'name': 'Eevee'},
+    103: {'hex': '67', 'decimal': '103', 'name': 'Flareon'},
+    104: {'hex': '68', 'decimal': '104', 'name': 'Jolteon', 'type': 'Electric'},
+    105: {'hex': '69', 'decimal': '105', 'name': 'Vaporeon'},
+    106: {'hex': '6A', 'decimal': '106', 'name': 'Machop'},
+    107: {'hex': '6B', 'decimal': '107', 'name': 'Zubat'},
+    108: {'hex': '6C', 'decimal': '108', 'name': 'Ekans'},
+    109: {'hex': '6D', 'decimal': '109', 'name': 'Paras', 'type': 'Bug'},
+    110: {'hex': '6E', 'decimal': '110', 'name': 'Poliwhirl'},
+    111: {'hex': '6F', 'decimal': '111', 'name': 'Poliwrath'},
+    112: {'hex': '70', 'decimal': '112', 'name': 'Weedle', 'type': 'Bug'},
+    113: {'hex': '71', 'decimal': '113', 'name': 'Kakuna', 'type': 'Bug'},
+    114: {'hex': '72', 'decimal': '114', 'name': 'Beedrill', 'type': 'Bug'},
+    115: {'hex': '73', 'decimal': '115', 'name': 'MissingNo. (Houndoom)'},
+    116: {'hex': '74', 'decimal': '116', 'name': 'Dodrio'},
+    117: {'hex': '75', 'decimal': '117', 'name': 'Primeape'},
+    118: {'hex': '76', 'decimal': '118', 'name': 'Dugtrio'},
+    119: {'hex': '77', 'decimal': '119', 'name': 'Venomoth', 'type': 'Bug'},
+    120: {'hex': '78', 'decimal': '120', 'name': 'Dewgong', 'type': 'Ice'},
+    121: {'hex': '79', 'decimal': '121', 'name': 'MissingNo. (Kingdra)'},
+    122: {'hex': '7A', 'decimal': '122', 'name': 'MissingNo. (Phanpy)'},
+    123: {'hex': '7B', 'decimal': '123', 'name': 'Caterpie', 'type': 'Bug'},
+    124: {'hex': '7C', 'decimal': '124', 'name': 'Metapod', 'type': 'Bug'},
+    125: {'hex': '7D', 'decimal': '125', 'name': 'Butterfree', 'type': 'Bug'},
+    126: {'hex': '7E', 'decimal': '126', 'name': 'Machamp'},
+    127: {'hex': '7F', 'decimal': '127', 'name': 'MissingNo. (Donphan)'},
+    128: {'hex': '80', 'decimal': '128', 'name': 'Golduck'},
+    129: {'hex': '81', 'decimal': '129', 'name': 'Hypno'},
+    130: {'hex': '82', 'decimal': '130', 'name': 'Golbat'},
+    131: {'hex': '83', 'decimal': '131', 'name': 'Mewtwo'},
+    132: {'hex': '84', 'decimal': '132', 'name': 'Snorlax'},
+    133: {'hex': '85', 'decimal': '133', 'name': 'Magikarp'},
+    134: {'hex': '86', 'decimal': '134', 'name': 'MissingNo. (Porygon2)'},
+    135: {'hex': '87', 'decimal': '135', 'name': 'MissingNo. (Stantler)'},
+    136: {'hex': '88', 'decimal': '136', 'name': 'Muk'},
+    137: {'hex': '89', 'decimal': '137', 'name': 'MissingNo. (Smeargle)'},
+    138: {'hex': '8A', 'decimal': '138', 'name': 'Kingler'},
+    139: {'hex': '8B', 'decimal': '139', 'name': 'Cloyster'},
+    141: {'hex': '8D', 'decimal': '141', 'name': 'Electrode'},
+    142: {'hex': '8E', 'decimal': '142', 'name': 'Clefable'},
+    143: {'hex': '8F', 'decimal': '143', 'name': 'Weezing'},
+    144: {'hex': '90', 'decimal': '144', 'name': 'Persian'},
+    145: {'hex': '91', 'decimal': '145', 'name': 'Marowak'},
+    147: {'hex': '93', 'decimal': '147', 'name': 'Haunter'},
+    148: {'hex': '94', 'decimal': '148', 'name': 'Abra'},
+    149: {'hex': '95', 'decimal': '149', 'name': 'Alakazam'},
+    150: {'hex': '96', 'decimal': '150', 'name': 'Pidgeotto'},
+    151: {'hex': '97', 'decimal': '151', 'name': 'Pidgeot'},
+    152: {'hex': '98', 'decimal': '152', 'name': 'Starmie'},
+    153: {'hex': '99', 'decimal': '153', 'name': 'Bulbasaur'},
+    154: {'hex': '9A', 'decimal': '154', 'name': 'Venusaur'},
+    155: {'hex': '9B', 'decimal': '155', 'name': 'Tentacruel'},
+    157: {'hex': '9D', 'decimal': '157', 'name': 'Goldeen'},
+    158: {'hex': '9E', 'decimal': '158', 'name': 'Seaking'},
+    163: {'hex': 'A3', 'decimal': '163', 'name': 'Ponyta'},
+    164: {'hex': 'A4', 'decimal': '164', 'name': 'Rapidash'},
+    165: {'hex': 'A5', 'decimal': '165', 'name': 'Rattata'},
+    166: {'hex': 'A6', 'decimal': '166', 'name': 'Raticate'},
+    167: {'hex': 'A7', 'decimal': '167', 'name': 'Nidorino'},
+    168: {'hex': 'A8', 'decimal': '168', 'name': 'Nidorina'},
+    169: {'hex': 'A9', 'decimal': '169', 'name': 'Geodude'},
+    170: {'hex': 'AA', 'decimal': '170', 'name': 'Porygon'},
+    171: {'hex': 'AB', 'decimal': '171', 'name': 'Aerodactyl'},
+    173: {'hex': 'AD', 'decimal': '173', 'name': 'Magnemite'},
+    176: {'hex': 'B0', 'decimal': '176', 'name': 'Charmander'},
+    177: {'hex': 'B1', 'decimal': '177', 'name': 'Squirtle'},
+    178: {'hex': 'B2', 'decimal': '178', 'name': 'Charmeleon'},
+    179: {'hex': 'B3', 'decimal': '179', 'name': 'Wartortle'},
+    180: {'hex': 'B4', 'decimal': '180', 'name': 'Charizard'},
+    185: {'hex': 'B9', 'decimal': '185', 'name': 'Oddish'},
+    186: {'hex': 'BA', 'decimal': '186', 'name': 'Gloom'},
+    187: {'hex': 'BB', 'decimal': '187', 'name': 'Vileplume'},
+    188: {'hex': 'BC', 'decimal': '188', 'name': 'Bellsprout'},
+    189: {'hex': 'BD', 'decimal': '189', 'name': 'Weepinbell'},
+    190: {'hex': 'BE', 'decimal': '190', 'name': 'Victreebel'}
+    }
 
 POKE = [0xD16B, 0xD197, 0xD1C3, 0xD1EF, 0xD21B, 0xD247] # - Pokémon (Again)
 STATUS = [0xD16F, 0xD19B, 0xD1C7, 0xD1F3, 0xD21F, 0xD24B] # - Status (Poisoned, Paralyzed, etc.)
@@ -572,21 +692,21 @@ MOVE2 = [0xD174, 0xD1A0, 0xD1CC, 0xD1F8, 0xD224, 0xD250]
 MOVE3 = [0xD175, 0xD1A1, 0xD1CD, 0xD1F9, 0xD225, 0xD251]
 MOVE4 = [0xD176, 0xD1A2, 0xD1CE, 0xD1FA, 0xD226, 0xD252]
 
-def pokemon_l(game):
-    pokemon_info = [{"slot": str(i + 1), "name": "", "level": "0", "moves": []} for i in range(6)]
-    for i in range(6):
-        p, l = game.get_memory_value(POKE[i]), game.get_memory_value(LEVEL[i])
-        hex_value = hex(int(p))[2:].upper()
-        matching_pokemon = next((entry for entry in pokemon_data if entry.get('hex') == hex_value), None)
-        if matching_pokemon:
-            pokemon_info[i]["name"] = matching_pokemon["name"]
-            pokemon_info[i]["level"] = str(l)
-            moves_addresses = [MOVE1[i], MOVE2[i], MOVE3[i], MOVE4[i]]
-            pokemon_info[i]["moves"] = []
-            for moves_address in moves_addresses:
-                move_value = game.get_memory_value(moves_address)
-                if move_value != 0x00:
-                    move_info = moves_dict.get(move_value, {})
-                    move_name = move_info.get("Move", "")
-                    pokemon_info[i]["moves"].append(move_name)
-    return pokemon_info
+def logs(game):
+    data_dict = {}
+    for i in range(len(POKE)): # 6 
+        p = game.get_memory_value(POKE[i])
+        l = game.get_memory_value(LEVEL[i])
+        moves_addresses = [MOVE1[i], MOVE2[i], MOVE3[i], MOVE4[i]]
+        n = poke_dict.get(p, {})
+        name = n.get('name', '')
+        slot_data = {'name': name,'level': l,'moves': []}
+        for moves_address in moves_addresses:
+            move_value = game.get_memory_value(moves_address)
+            if move_value != 0x00:
+                move_info = moves_dict.get(move_value, {})
+                move_name = move_info.get("Move", "")
+                slot_data["moves"].append(move_name)
+        data_dict[f'slot{i+1}'] = slot_data
+
+    return data_dict
